requirements.txt
urllib3~=1.24.1
Pillow~=7.0.0